#27 Includes refactoring
