+++ src/parser.c
@@ -1,0 +1,13 @@
+// Filename: parser.c
+// IFJ 2022/2023 project
+// TRP variant
+//
+// Parser for syntax analysis
+//
+// Author: <xsklen14> - David Sklenář
+// Author: <xproch0o> - Marek Procházka
+// Author: <xzavad20> - Lukáš Zavadil
+// Author: <xsvetl07> - Adam Světlík
+
+// LOCAL INCLUDES
+#include "parser.h"
