+++ src/parser.h
@@ -1,0 +1,19 @@
+#pragma once
+
+// Filename: parser.h
+// IFJ 2022/2023 project
+// TRP variant
+//
+// Parser for syntax analysis - header file
+//
+// Author: <xsklen14> - David Sklenář
+// Author: <xproch0o> - Marek Procházka
+// Author: <xzavad20> - Lukáš Zavadil
+// Author: <xsvetl07> - Adam Světlík
+
+// EXTERNAL INCLUDES
+#include <stdio.h>
+#include <stdlib.h>
+#include <stdbool.h>
+
+// LOCAL INCLUDES
