+++ src/symtable.c
@@ -1,0 +1,13 @@
+// Filename: symtable.c
+// IFJ 2022/2023 project
+// TRP variant
+//
+// Symtable
+//
+// Author: <xsklen14> - David Sklenář
+// Author: <xproch0o> - Marek Procházka
+// Author: <xzavad20> - Lukáš Zavadil
+// Author: <xsvetl07> - Adam Světlík
+
+// LOCAL INCLUDES
+#include "symtable.h"
